data-generator-polygon/pyproject.toml
[project]
name = "data-generator-polygon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "polygon-api-client>=1.14.3",
    "psutil>=6.1.1",
    "requests>=2.32.3",
]
